refactor(docs): dry markdown wrapping components and shink document styling
