+++ .storybook/components/Docs/GettingStarted/GettingStarted.stories.tsx
@@ -1,30 +1,22 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { StoryObj } from '@storybook/react';
 import React from 'react';
-import { GettingStartedDocumentation } from './GettingStartedDocumentation';
 import GettingStartedDocs from '../../../../docs/GETTING_STARTED.md';
-import styles from '../markdown.module.css';
+import { Documentation } from '../Documentation';
 export default {
   title: 'Getting Started',
-  component: GettingStartedDocumentation,
+  component: GettingStartedDocs,
   parameters: {
     chromatic: { disableSnapshot: true },
     axe: {
       skip: true,
     },
   },
-  decorators: [
-    (Story) => (
-      <div className={styles['markdown']}>
-        <Story />
-      </div>
-    ),
-  ],
-} as Meta;
+};
 
-type Args = React.ComponentProps<typeof GettingStartedDocumentation>;
-
-export const GettingStarted: StoryObj<Args> = {
-  args: {
-    children: <GettingStartedDocs />,
-  },
+export const GettingStarted: StoryObj = {
+  render: () => (
+    <Documentation>
+      <GettingStartedDocs />
+    </Documentation>
+  ),
 };
